refactor(WebsocketUserService): replace strings with proper variables and move some files to their appropriate locations
