+++ backend/burning-okr/burning-okr-app/src/test/java/org/burningokr/config/WebSocketConfigTest.java
@@ -1,5 +1,6 @@
 package org.burningokr.config;
 
+import org.burningokr.service.security.websocket.WebSocketAuthentication;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
