+++ backend/burning-okr/burning-okr-app/src/test/java/org/burningokr/service/WebsocketUserServiceTest.java
@@ -1,7 +1,7 @@
 package org.burningokr.service;
 
-import org.burningokr.config.WebSocketAuthentication;
 import org.burningokr.model.users.User;
+import org.burningokr.service.security.websocket.WebSocketAuthentication;
 import org.burningokr.service.userhandling.UserService;
 import org.burningokr.service.websocket.WebsocketUserService;
 import org.junit.jupiter.api.Assertions;
