+++ backend/burning-okr/burning-okr-logic/src/main/java/org/burningokr/service/exceptions/AuthorizationHeaderException.java
@@ -1,4 +1,4 @@
-package org.burningokr.exceptions;
+package org.burningokr.service.exceptions;
 
 public class AuthorizationHeaderException extends RuntimeException {
     public AuthorizationHeaderException(String message) {
