+++ backend/burning-okr/burning-okr-logic/src/main/java/org/burningokr/service/websocket/WebsocketUserService.java
@@ -3,6 +3,7 @@ package org.burningokr.service.websocket;
 import lombok.RequiredArgsConstructor;
 import org.burningokr.model.monitoring.UserId;
 import org.burningokr.model.users.User;
+import org.burningokr.service.security.websocket.WebSocketAuthentication;
 import org.burningokr.service.userhandling.UserService;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.stereotype.Service;
@@ -16,10 +17,10 @@ public class WebsocketUserService {
 
   public User findByAccessor(StompHeaderAccessor stompHeaderAccessor) {
     Map<String, Object> sessionAttributes = stompHeaderAccessor.getSessionAttributes();
-    if (sessionAttributes == null || !sessionAttributes.containsKey("userId")) {
+    if (sessionAttributes == null || !sessionAttributes.containsKey(WebSocketAuthentication.USER_SESSION_ATTRIBUTE_KEY)) {
       throw new RuntimeException("StompHeaderAccessor does not contain (valid) userId.");
     }
-    UserId userId = new UserId(sessionAttributes.get("userId").toString());
+    UserId userId = new UserId(sessionAttributes.get(WebSocketAuthentication.USER_SESSION_ATTRIBUTE_KEY).toString());
     return userService.findById(userId.getUserId()).orElseThrow();
   }
 }
